feat: update vad rust
